% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ.R
\name{EOQ}
\alias{EOQ}
\title{Deterministic Economic Order Quantity Model}
\usage{
EOQ(d, A, h, p = 0, b1 = 0)
}
\arguments{
\item{{d}}{{demand rate}}

\item{{A}}{{fixed order cost}}

\item{{h}}{{holding cost}}

\item{{p}}{{production rate, default value 0}}

\item{{b1}}{{backorder cost, default value 0}}
}
\value{
A list containing:
\itemize{
\item {Q:} {the optimal order quantity}
\item {C:} {the optimal cost}
\item {x:} {is returned only in the case of the backorder model, the fraction of the demand being backorderd }
\item or NULL values, in case of inadequate input.
}
}
\description{
It implements three single period, single echelon, economic order quantity (EOQ)
model variants:
the classic, with production,and with planned backorders.
See chapter 4 in Axsäter's book for the details
}
\details{
The appropriate model is automatically selected based on the user input.
The default model is the classic EOQ and corresponds to the absence (i.e., zero values)
of production rate and backorder cost.
}
\examples{
\dontrun{** Example (based on Examples 4.1 and 4.3 in Axsäter) **
An item costs $100 per unit and has a holding cost rate per year corresponding to 20\% of
this value. The constant demand is 300 units per year. The ordering cost is $200. Clearly,
h= 0.20(100) = $20 per year.}

EOQ(d=300,A=200,h=20)

\dontrun{Assume that bacorders are allowed with a backorder cost $100 per unit and year.
 What would then be the order quantity? What the fraction of bacorded demand?}

 EOQ(d=300,A=200,h=20, b1=100)

\dontrun{Assume instead that the item is produced at a rate of 330 units per year.
 What would then be the order quantity?}

EOQ(d=300,A=200,h=20, p=330)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
EOQsensitivity, DEOQ, IDEOQ, EOQ_visualize_costs
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
