% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQ_visualize_costs.R
\name{EOQ_visualize_costs}
\alias{EOQ_visualize_costs}
\title{Visualization of the costs involved in the EOQ model}
\usage{
EOQ_visualize_costs(A = 10, d = 1500, h = 0.65, Q = seq(30, 300, by = 10))
}
\arguments{
\item{{d}}{{Yearly demand rate, default value 1500}}

\item{{A}}{{Fixed ordering cost, default value 10}}

\item{{h}}{{Yearly holding cost per unit, default value 0.65}}

\item{{Q}}{{a vector containing range of demands around EOQ. Provide an empty vector c() to let
the automatic generation of a range. Default range is 30:300 by a step 10}}
}
\value{
A list consisting of two elements:
\itemize{
\item {Plot} {contains the plot of the costs}
\item {Table} {is data frame of four columns:\itemize{
\item{Q} {the range of order quantities}
\item{HC} {the corresponding range of holding costs}
\item{OC} {the corresonding range of order costs}
\item{TC} {the corresponding range of total costs}}}
\item or NULL, if supplied with inadequate holding cost (<=0)
}
}
\description{
A plot of the ordering, holding and total costs is produced for a range
of order quantities provided by the user. The optimal order quantity (EOQ)
is computed and plotted too.
}
\examples{
\dontrun{** Example 1 **
Run without arguments for the default plot and default output, like this:}

EOQ_visualize_costs()

\dontrun{** Example 2 **
Store the result in a list and view it using print(), like this:}

res <- EOQ_visualize_costs()
print( res$Plot )
print( res$Table )

\dontrun{** Example 3 **
You can supply your own data, like this:}

res <- EOQ_visualize_costs(A=100, d=300, h=20, Q=seq(20,100,by=5))

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
EOQsensitivity, DEOQ, IDEOQ, EOQ
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
