% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQdisc.R
\name{EOQdisc}
\alias{EOQdisc}
\title{EOQ model with order quantity (all units) discounts}
\usage{
EOQdisc(d, A, Qb, v, r = 0, h0 = 0, h = 0)
}
\arguments{
\item{{d}}{{Yearly demand rate}}

\item{{A}}{{Fixed ordering cost}}

\item{{Qb}}{{A vector of quantity break points}}

\item{{v}}{{A vector of unit purchase price corresponding to quantity break points}}

\item{{r}}{{Interest rate for holding a unit as a percentage of purchase price. Default value is 0}}

\item{{h0}}{{Capital cost per unit and year. Default value is 0}}

\item{{h}}{{Yearly holding cost per unit. Default value is 0}}
}
\value{
A list consisting of two elements:
\itemize{
\item{Sol} {a vector containing the optimal solution in two elements:
\itemize{
\item{minQ} {the optimal quantity that minimizes total cost}
\item{minC} {the optimal total cost}}}
\item{Table} {a data frame with two columns containing intermediate result:
\itemize{
\item{Q} {feasible quantities}
\item{C} {corresponding costs}}} }
}
\description{
This model is based on the assumption that the supplier offers
discounts based on the quantity ordered. The all units discount
case is considered here, that is, given the quantity break points for
the different prices, all units to the right of a breakpoint incur
the price determined by that breakpoint.
}
\details{
Note that it is expected that either r (and h0) are supplied or h. If h is not supplied, i.e,
left to the default value 0, then it is set to h0+r*v. h0 may be left to its default 0-value when supplying r.
Thus, to have a meaningful model at least r or h must be supplied.
}
\examples{
## ** Example 1 (based on Axsäter's Example 4.2) **
## Suppose a supplier offers an item for a unit price of $100 unless the order quantity is 100
## or more units in which case the unit price is $95. Assume that the capital cost per unit and year
## is $5 and that the interest rate is 0.2. The yearly demand for the item is constant
## 300 units. The order cost is $200.

EOQdisc(d=300, A=200, Qb=c(100), r=0.2, v=c(100,95), h0=5)

## ** Example 2 ( based on Snyder-Shen's Example 3.6) **
## A candy with yearly demand of 1300 units is ordered for a fix order cost of $8
## from a supplier who offers it for a unit price of $0.75. However,  if 400 or more up
## to 799 are purchased, the unit price is discounted to 0.72; and if 800 or more are
## purchased, the price is further discounted to $0.68 per unit. The interest rate is 0.3
## of the purchased price per unit per year.

EOQdisc(d=1300, A=8, Qb=c(400,800),r=0.3, v=c(0.75,0.72,0.68) )

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
}
}
\seealso{
EOQsensitivity, EOQ_visualize_costs, EOQincr, EOQ
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
