% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOQsensitivity.R
\name{EOQsensitivity}
\alias{EOQsensitivity}
\title{Sensitivity of the Economic Order Quantity}
\usage{
EOQsensitivity(Qopt, Qother)
}
\arguments{
\item{{Qopt}}{{the optimal order quantity}}

\item{{Qother}}{{a tentative order quantity}}
}
\value{
A numerical value corresponding to the cost quotient, or NULL, in
case of inadequate input.
}
\description{
Computes the quotient (C/C*) of the cost (C) for a non-optimal versus
the cost (C*) for the optimal order quantity.
}
\examples{
EOQsensitivity(Qopt=200,Qother=300)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
EOQ, DEOQ, IDEOQ, EOQ_visualize_costs
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
