% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEOQ.R
\name{IDEOQ}
\alias{IDEOQ}
\title{EOQ model with order quantity (incrimental) discounts}
\usage{
IDEOQ(d, A, Qb, v, r)
}
\arguments{
\item{{d}}{{Yearly demand rate}}

\item{{A}}{{Fixed ordering cost}}

\item{{Qb}}{{A vector of quantity break points}}

\item{{v}}{{A vector of unit purchase price corresponding to quantity break points}}

\item{{r}}{{Interest rate for holding a unit as a percentage of purchase price}}
}
\value{
A list consisting of two elements:
\itemize{
\item{Sol} {a vector containing the optimal solution in two elements:
\itemize{
\item{optQ} {the optimal quantity that minimizes total cost}
\item{optC} {the optimal total cost}}}
\item{Table} {a data frame with three columns containing intermediate result:
\itemize{
\item{Q} {quantities}
\item{C} {corresponding costs}
\item{Feas} {provides indication of feasibility}}} }
}
\description{
This model is based on the assumption that the supplier offers
discounts based on the quantity ordered. The incremental discount
case is considered here, that is, given the quantity break points for
the different prices, the units within an interval with end points defined by
consecutive breakpoints incur the purchase price for that interval.
}
\examples{
## ** Example 1 (based on Axsäter's Example 4.2) **
## Suppose a supplier offers an item for a unit price of $100 unless the order quantity is 100
## or more units in which case the unit price is $95. Assume that the interest rate is 0.2.
## The yearly demand for the item is constant 300 units. The order cost is $200.

IDEOQ(d=300, A=200, Qb=c(100), r=0.2, v=c(100,95) )

## ** Example 2 ( based on Snyder-Shen's Example 3.7) **
## A candy with yearly demand of 1300 units is ordered for a fix order cost of $8
## from a supplier who offers it for a unit price of $0.75. However,  if 400 or more up
## to 799 are purchased, the unit price is discounted to 0.72; and if 800 or more are
## purchased, the price is further discounted to $0.68 per unit. The interest rate is 0.3
## of the purchased price per unit per year.

IDEOQ(d=1300, A=8, Qb=c(400,800),r=0.3, v=c(0.75,0.72,0.68) )

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
}
}
\seealso{
EOQsensitivity, EOQ_visualize_costs, DEOQ, EOQ
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
