% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Indicators.R
\name{Inventory_Indicators}
\alias{Inventory_Indicators}
\alias{IL}
\alias{IP}
\alias{EILm}
\title{Functions to evaluate inventory indicators such as level and position}
\usage{
IL(mu, Q, R)

IP(Q, R)

EILm(mu, sigma, L, Q, R)
}
\arguments{
\item{{mu}}{{mean value of demand}}

\item{{R}}{{reorder point}}

\item{{Q}}{{order quantity}}

\item{{L}}{{lead time}}
}
\value{
A single numeric value. NULL if wrong input.
}
\description{
Functions to compute inventory indicators such as level and position
}
\details{
Function IL(): Computets the average inventory level

Function IP(): Computes the average  inventory position

Function EILm(): Computes the expected amount of backorders
}
\examples{
IL(mu=100, Q=85, R=90)
IP(Q=100, R=90)
EILm(mu=150,sigma=40,L=4,Q=100, R=95)
}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
