% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointRQb1.R
\name{JointRQb1}
\alias{JointRQb1}
\title{Stochastic model for joint optimal determination of R and Q}
\usage{
JointRQb1(mu, sigma, L, h, b1, A, Qo, Ro)
}
\arguments{
\item{mu:}{mean value of demand}

\item{sigma:}{standard deviation of demand}

\item{L:}{lead-time}

\item{h:}{holding cost}

\item{b1:}{stockout cost}

\item{A:}{fix order cost}

\item{Qo:}{initial guess of order quantity}

\item{Ro:}{initial guess of reorder point}
}
\value{
A list containing:
\itemize{
\item{Startpoint:}  {a vector with elements Qo, Ro as above}
\item{Startcost:}   {the cost Co for these guesses}
\item{Finalpoint:}  {a vector with elements Q*, R*}
\item{Finalcost:}   {the cost C* for these final values}
\item{Optimum:}     {TRUE if optimality of final point is verified, FALSE otherwise
(in case of FALSE, the final point can be used as a new start point)}
}
}
\description{
Stochastic model for joint optimal determination of R and Q
}
\details{
It implements a stochastic model for joint determination of R and Q
with normal distribution of demand and stock out cost b1 per unit
and time unit. The model is that of section 6.1.2 in Axsäter's book (see
references below)

The Nelder-Mead method from the package "optimx" is applied for
direct minimization of the expected total cost.
}
\examples{
\dontrun{
#Example 6.1 in Axsäter's book, page 111#
#Suppose that the  demand per time unit is normally
#distributed with μ = 50 and σ = 20. The costs are A = 100, h = 2, and b1 = 20.
#Suppose that the lead-time is L = 4. The order quantity Q and the reorder point
#R are sought.
#
# We need to provide initial guesses for R and Q. Thus, let Ro=224.76 and Qo= 70.71.
}

JointRQb1(mu=50, sigma=20, L=4, h=2, b1=20, A=100, Qo=70.71, Ro=224.76)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
JointRQu, JointRQb1iter, JointRQss, JointRQhw, JointRQb1S2
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
