% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointRQb1iter.R
\name{JointRQb1iter}
\alias{JointRQb1iter}
\title{Stochastic model for joint optimal determination of R and Q}
\usage{
JointRQb1iter(mu, sigma, L, h, b1, A)
}
\arguments{
\item{mu:}{mean value of demand}

\item{sigma:}{standard deviation of demand}

\item{L:}{lead-time}

\item{h:}{holding cost}

\item{b1:}{stockout cost}

\item{A:}{fix order cost}
}
\value{
A data frame containing the results of the iterations in three columns:
\itemize{
\item{Q:}  {the successive order quantities generated}
\item{R:}   {the successive reorder points generated}
\item{C:}  {the corresponding total cost for each successive pair of Q and R}
}
}
\description{
Stochastic model for joint optimal determination of R and Q
}
\details{
It implements a stochastic model for joint determination of R and Q
with normal distribution of demand and stock out cost b1 per unit
and time unit. The model is that of section 6.1.2 in Axsäter's book (see
references below)

The iterative method of alternatingly solving the first order optimality
equations is applied in order to minimize the expected total cost.
The function "uniroot" of the package "stats" is used in order to solve
the equation for R.

For more details on the iterative approach see section 6.1.2 in Axsäter's book.
}
\examples{
\dontrun{
#Example 6.1 in Axsäter's book, page 111#
#Suppose that the  demand per time unit is normally
#distributed with μ = 50 and σ = 20. The costs are A = 100, h = 2, and b1 = 20.
#Suppose that the lead-time is L = 4. The order quantity Q and the reorder point
#R are sought.
#
# No initial guesses are required since the algorithm will start by utilizing
# the EOQ produced order quantity.}

JointRQb1iter(mu=50, sigma=20, L=4, h=2, b1=20, A=100)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
JointRQb1, JointRQu, JointRQss, JointRQhw, JointRQb1S2
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
