% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointRQb1S2.R
\name{JointRQb1S2}
\alias{JointRQb1S2}
\title{Stochastic model for joint optimal determination of R and Q for given service level S2}
\usage{
JointRQb1S2(mu, sigma, L, h, b1, A, S2, Qo, Ro)
}
\arguments{
\item{mu:}{mean value of demand}

\item{sigma:}{standard deviation of demand}

\item{L:}{lead-time}

\item{h:}{holding cost}

\item{b1:}{stockout cost}

\item{A:}{fix order cost}

\item{S2:}{service level of type II}

\item{Qo:}{initial guess of order quantity}

\item{Ro:}{initial guess of reorder point}
}
\value{
A list containing:
\itemize{
\item{Startpoint:}  {a vector with elements Qo, Ro as above}
\item{Startcost:}   {the cost Co for these guesses}
\item{Finalpoint:}  {a vector with elements Q*, R*}
\item{Finalcost:}   {the cost C* for these final values}
\item{Optimum:}     {TRUE if optimality of final point is verified, FALSE otherwise
(in case of FALSE, the final point can be used as a new start point)}
}
}
\description{
Stochastic model for joint optimal determination of R and Q for given service level S2
}
\details{
It implements a stochastic model for joint determination of R and Q
with normal distribution of demand and stock out cost b1 per unit
and time unit under constraint for service level of type II (S2).
The model is that of section 6.1.3 in Axsäter's book (see
references below)

A Nonlinear Augmented Lagrangian SQP Based Approach, "solnp()" from the
package "Rsolnp" is applied for direct minimization of the expected total cost.
}
\examples{
\dontrun{
#Example 6.2 in Axsäter's book, page 113#
#Suppose that the  demand per time unit is normally
#distributed with μ = 50 and σ = 20. The costs are A = 100, h = 2, and b1 = 0.
#Suppose that the lead-time is L = 4 and the service level of type II requested at 90\%.
#The order quantity Q and the reorder point R are sought.
#
# We need to provide initial guesses for R and Q. Thus, let Ro=224.76 and Qo= 70.71.
#}

JointRQb1S2(mu=50, sigma=20, L=4, h=2, b1=0, A=100, S2=0.9, Qo=70.71, Ro=224.76)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
JointRQb1, JointRQb1iter, JointRQss, JointRQhw, JointRQu
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
