% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointRQss.R
\name{JointRQss}
\alias{JointRQss}
\title{Stochastic model for joint optimal determination of R and Q}
\usage{
JointRQss(A, h, b1, mu, sigma, L, maxiter = 20, eps = 5e-08)
}
\arguments{
\item{mu:}{mean value of demand}

\item{sigma:}{standard deviation of demand}

\item{L:}{lead-time}

\item{h:}{holding cost}

\item{b1:}{stockout cost}

\item{A:}{fix order cost}

\item{maxiter:}{maximum number of iterations, default 20}

\item{eps:}{size of accepted relative difference between two consecutive values}
}
\value{
Returns a list containing:
\itemize{
\item{Q:}  {order quantity}
\item{R:}   {reorder point}
\item{ExpCost:}  {expected total cost}
\item{Iterations:} {the number of iterations executed. If 0 iterations returned,  check solution for feasibility.}
}
}
\description{
Stochastic model for joint optimal determination of R and Q
}
\details{
It implements a stochastic model for joint determination of R and Q
with normal distribution of demand and stock out cost b1 per unit.

Model Assumptions:
\enumerate{
\item Demand with yearly mean mu (μ) and standard deviation sigma (σ)
\item Stockout cost b1 per unit
\item Lead-time demand normally distributed with mean mu (μ') and standard deviation sigma (σ')
}

The model  (see section 4.3.2 in Snyder and Shen's book) is similar to that of
section 6.1.2 in Axsäter's book.

The iterative method of alternatingly update order quantity and reorder point
is applied in order to minimize the expected total cost.
}
\examples{
\dontrun{
#Example #
#Suppose that the mean value of demand is normally distributed with μ = 50 and σ=20.
#The lead-time is 4 weeks. The costs are A = 100, h = 2, and b1 = 20.
#The order quantity Q and the reorder point R are sought.
#}

JointRQss(mu=50, sigma=20, h=2, b1=20, A=100, L=4, maxiter=10)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J. M. (2011) Fundamentals of Supply Chain Theory, Wiley
}
}
\seealso{
JointRQb1, JointRQb1iter, JointRQu, JointRQhw, JointRQb1S2
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
