% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointRQu.R
\name{JointRQu}
\alias{JointRQu}
\title{Stochastic model for joint optimal determination of R and Q}
\usage{
JointRQu(A, h, b1, mu, a = 0, b = 100, maxiter = 20, eps = 5e-06)
}
\arguments{
\item{a, b:}{interval of uniform distribution for lead-time demand, default (0,100)}

\item{mu:}{mean value of year demand}

\item{h:}{holding cost}

\item{b1:}{stockout cost}

\item{A:}{fix order cost}

\item{maxiter:}{maximum number of iterations, default 20}

\item{eps:}{size of accepted relative difference between two consecutive values}
}
\value{
Returns a list containing:
\itemize{
\item{Q:}  {order quantity}
\item{R:}   {reorder point}
\item{ExpCost:}  {expected total cost}
\item{Iterations:} {the number of iterations executed. If 0 iterations returned,  check solution for feasibility.}
}
}
\description{
Stochastic model for joint optimal determination of R and Q
}
\details{
It implements a stochastic model for joint determination of R and Q
with uniform distribution of lead-time demand. The model is similar but not equivalent to that of
section 6.1.2 in Axsäter's book (see references below).

The iterative method of alternatingly update order quantity and reorder point
is applied in order to minimize the expected total cost.
}
\examples{
\dontrun{
#Example #
#Suppose that the mean value of year demand is μ = 50, and that the lead-time demand
#is uniformly distrubted between a lower limit a=10 and an upper limit b=65.
#The costs are A = 100, h = 2, and b1 = 20.
#The order quantity Q and the reorder point
#R are sought.
#}

JointRQu(mu=50, h=2, b1=20, A=100, a=10, b=65, maxiter=10)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
JointRQb1, JointRQb1iter, JointRQss, JointRQhw, JointRQb1S2
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
