% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRiter.R
\name{JRiter}
\alias{JRiter}
\title{Joint Replenishment - Iterative Approach}
\usage{
JRiter(d, A, h, a)
}
\arguments{
\item{{d}}{{demand per time unit for each item}}

\item{{A}}{{set up cost for the group}}

\item{{h}}{{holding cost per item unit and time unit}}

\item{{a}}{{set up cost per item}}
}
\value{
A list containing
\itemize{
\item{BasicPeriod} {corresponds to cycle time for the "first" item}
\item{Multipliers} {of the basic period to express cycle
times for the rest of the items}
\item{Ordered} {item indeces ordered}
\item{LBD} {cost lower bound}
\item{UBD} {cost upper bound}
}
}
\description{
A deterministic joint replenishment model is considered in which
there is a joint set up cost independent of the number of items ordered
olus individual set up costs for each item. Cycle times (and, consequently,
batch quantities) are assumed constant. No backorders are allowed and the
production time is being disregarded. The problem is to determine the cycle
times that minimize the sum of holding and setup costs.
See section 7.3  in Axs채ter's book for more information.

The iterative approach (see section 7.3.1.1 in Axs채ter) is a heuristic algorithm
which attempts to determine cycle times as integer multiples of the smallest
cycle time, initially approximated with the EOQ lot size formula.
}
\examples{
## ** Example based on Example 7.2 in Axs채ter **
## Four item with joint set up cost $300 and equal individual set up
## and holding costs of $50 and $10 respectively. The demand for the 
## four items are 5000, 1000, 700 and 100 respectively.
##
JRiter(d=c(5000,1000,700,100),A=300,h=c(10,10,10,10),a=c(50,50,50,50))


}
\references{
\itemize{
\item Axs채ter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
JRroundy
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
