% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardized_Functions.R
\name{Normal_Functions}
\alias{Normal_Functions}
\alias{G}
\alias{H}
\alias{phi}
\alias{Phi}
\alias{PhInv}
\alias{F}
\alias{f}
\title{Standardized normal distribution functions in Axsäter's notation}
\usage{
G(x)

H(x)

phi(x)

Phi(x)

PhInv(x)

F(x, Q, R, mu, sigma, L)

f(x, Q, R, mu, sigma, L)
}
\arguments{
\item{{x}}{{true function argument}}

\item{{Q}}{{parameter representing order quantity}}

\item{{R}}{{parameter representing reorder point}}

\item{{mu}}{{parameter representing demand mean}}

\item{{sigma}}{{parameter representing demand's standard deviation}}

\item{{L}}{{parameter representing lead-time}}
}
\value{
a numeric value or NULL in case of erroneous input
}
\description{
Standardized normal distribution functions in Axsäter's notation
}
\details{
The following functions are made available:
\enumerate{
\item Standardized Loss Function $G(x)=\\int_x^\\infty (v-x)\\phi(v)dv$.
Named G(x) after Axsäter's book, Chapter 5
\item The integral
$H(x) = \\int_x^\\infty G(v)dv$
of the standrardized loss function G(x)
as defined in Axsäter's book, Chapter 5.
\item Normal distribution functions in Axsäter's notation of Chapter 5:
}
\itemize{
\item Standardized normal density function phi(x), i.e., φ(x).
\item Standardized normal distribution function Phi(x), i.e.,  Φ(x).
\item Standardized normal inverse distribution function PhInv(x), i.e., $Phi^\{-1\}()$.
}
\enumerate{
\item Distribution and density functions of the inventory level in steady state (Axsäter's notation in Chapter 5)
}
\itemize{
\item Distribution function F(x) with additional input parameters Q, R, μ, σ, L.
\item Density function f(x) with additional input parameters Q, R, μ, σ, L.
}
}
\examples{
G(0.45)
H(0.55)
phi(0.58)
Phi(0.55)
PhInv(0.56)
F(x=0.15,Q=50,R=65,mu=68,sigma=25,L=2)
f(x=0.45, Q=50, R=65, mu=68, sigma=25, L=2)
}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
