% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_R_vs_S1.R
\name{Plot_R_vs_S1}
\alias{Plot_R_vs_S1}
\title{Plots Reorder Point (R) as function of Service Level Type 1 (S1)}
\usage{
Plot_R_vs_S1(S1low = 0.5, sigma, mu)
}
\value{
A list with two elements:
\itemize{
\item{\code{Plot}} {contains the plot. If stored, can be desplayed with \code{print()}}
\item{\code{Table}} {contains in four columns, \code{S1}, \code{R}, \code{SS} and \code{z}, the corresponding values for the specified range of service level 1.}
}
}
\description{
Plots Reorder Point (R), Safety Stock (SS) and Safety Factor (z) as functions
of Service Level Type 1 (S1). The purpose is to demonstrate the
nonlinearity of the relations, and particularly the accelation
in the increase that occurs as the 100\% limit of the service
level is approached.
}
\examples{
## ** Example 1 **
## For demand with mean 10 and standard deviation sqrt(10), leaving the lower limit
## of the service level to the default 50\%:
##
Plot_R_vs_S1(sigma=sqrt(10), mu=10)

## ** Example 2 **
## As above but specifying a lower limit of 30\%
##
Plot_R_vs_S1(S1low=0.3, sigma=sqrt(10), mu=10)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
Reorder_Point_S1, Reorder_Point_S2
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
