% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silver_Meal.R
\name{Silver_Meal}
\alias{Silver_Meal}
\title{Siver-Meal heuristic for lot-sizing with time-varying demand}
\usage{
Silver_Meal(d, A, h = 0, v = 0, r = 0)
}
\arguments{
\item{{d}}{{vector specifying the demands of the periods}}

\item{{A}}{{Fixed ordering cost}}

\item{{v}}{{Unit purchase price}}

\item{{r}}{{Interest rate for holding a unit as a percentage of purchase price. Default value is 0}}
}
\value{
A list consisting of
\itemize{
\item{\code{Costs}} {vector containing three elements:
\itemize{
\item{\code{TotalCost}} {the total policy cost}
\item{\code{TotalOrderCost}} {additive component of total cost}
\item{\code{TotalHoldingCost}} {additive component of total cost}}}
\item{\code{OrderPeriod}} {policy vector indicating for each period demand the order period}}
}
\description{
This model assumes that there is a finite number of periods and the demand for
each such period is known and realizable at the beginning of the period. It is
further assumed that there is no initial stock. The delivery of an entire order,
when it happens, deliverss the entire order quantity. Holding and ordering costs
are constant over time. No back orders are allowed.

The Silver-Meal algorithm is a heuristic alternative to the optimizing dynamic programming
algorithm by Wagner-Whitin. Although it can miss the optimal solution, it does often
produces it, however, since it does not verify optimality, you cannot be asured of it.
See sections 4.5, 4.6, and 4.7 in Axsäter's book.
}
\examples{
#' ## ** Example 1 (based on Example 4.5 in Axsäter) **
## The ordering cost is $300 and the holding cost $1 per unit and period.
## The horizon is 10 periods long and the demands for the 10 periods are:
## 50, 60, 90, 70, 30, 100, 60, 40, 80, 20
##

Silver_Meal(d=c(50, 60, 90, 70, 30, 100, 60, 40, 80, 20), A=300, h=1)

## ** Example 2 (based on Example 3.9 in Snyder-Shen) **
## Fixed ordering cost is $500, the holding cost $2 per unit and period.
## There are four periods with corresponding demand 90, 120, 80 and 70.
##

Silver_Meal(d=c(90, 120, 80, 70), A=500, h=2)

## ** Example 3 **
## As in Example 2 but the unit purchase price is $20 and the
## interest rate defining the holding cost per unit and period is 0.1.
##

Silver_Meal(d=c(90, 120, 80, 70), A=500, v=20, r=0.1)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
}
}
\seealso{
Wagner_Whitin, Balance_Costs
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
