% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wagner_Whitin.R
\name{Wagner_Whitin}
\alias{Wagner_Whitin}
\title{Wagner-Whitin algorithm for lot-sizing with time-varying demand}
\usage{
Wagner_Whitin(d, A, h = 0, v = 0, r = 0, recursion = c("forward", "backward"))
}
\arguments{
\item{{d}}{{vector specifying the demands of the periods}}

\item{{A}}{{Fixed ordering cost}}

\item{{v}}{{Unit purchase price}}

\item{{r}}{{Interest rate for holding a unit as a percentage of purchase price. Default value is 0}}

\item{{recursion}}{{A character constant (string), either "forward" or "backward"}}
}
\value{
A list consisting of
\itemize{
\item{\code{DPtable}} {a data frame with columns
\itemize{
\item{\code{DynPtable.k}} {where k=1..number of periods, containing the Dyn.P. cost calculations for each period}
\item{\code{IndexMin}} {identifying the column with the minimum cost on each row}}
}
\item{\code{optCost}} {the optimal ordering and holding cost}
\item{\code{totalCost}} {purchase or production cost added to optCost}
}
}
\description{
This model assumes that there is a finite number of periods and the demand for
each such period is known and realizable at the beginning of the period. It is
further assumed that there is no initial stock. The delivery of an entire order,
when it happens, deliverss the entire order quantity. Holding and ordering costs
are constant over time. No back orders are allowed.

The Wagner-Whitin algorithm is a clever realization of a dynamic programming approach
that utilizes the zero-inventory (ZIO) property inherent to the model. See sections
4.5 and 4.6 in Axsäter's book and section 3.3 in Snyder-Shen's book. The algorithm
seeks to derive the optimal order plan that minimizes the total cost, that is the
sum of ordering and holding costs (to which puchase cost may be added, if available).
}
\details{
\itemize{
\item The recursion specifies the course of computations in relation to the time periods.
\item The optimal policy is determined from \code{DPtable} with the support of the indices in \code{IndexMin}. Alternative
optimal policies may exist and are similarly identified.
\item Total purchase cost is fixed since the total demand is fixed, consequently it has no effect on the optimization and
is typically ignored.
\item If there is no purchase cost, then \code{optCost = totalCost}.
}
}
\examples{
## ** Example 1 (based on Example 4.5 in Axsäter) **
## The ordering cost is $300 and the holding cost $1 per unit and period.
## The horizon is 10 periods long and the demands for the 10 periods are:
## 50, 60, 90, 70, 30, 100, 60, 40, 80, 20
##
## Forward recursion (default):

Wagner_Whitin(d=c(50, 60, 90, 70, 30, 100, 60, 40, 80, 20), A=300, h=1)

## Backward recursion:

Wagner_Whitin(d=c(50, 60, 90, 70, 30, 100, 60, 40, 80, 20), A=300, h=1, recursion="backward")

## ** Example 2 (based on Example 3.9 in Snyder-Shen) **
## Fixed ordering cost is $500, the holding cost $2 per unit and period.
## There are four periods with corresponding demand 90, 120, 80 and 70.
##
## Forward recursion (default):

Wagner_Whitin(d=c(90, 120, 80, 70), A=500, h=2)

## Backward recursion:

Wagner_Whitin(d=c(90, 120, 80, 70), A=500, h=2, recursion="backward")

## ** Example 3 **
## As in Example 2 but the unit purchase price is $20 and the
## interest rate defining the holding cost per unit and period is 0.1.
##
## Forward recursion (default):

Wagner_Whitin(d=c(90, 120, 80, 70), A=500, v=20, r=0.1)

## Backward recursion:

Wagner_Whitin(d=c(90, 120, 80, 70), A=500, v=20, r=0.1, recursion="backward")

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
}
}
\seealso{
Silver_Meal, Balance_Costs
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
